copy methods generation; hashmapstorage

+++ src/main/java/com/github/kaeluka/cflat/storage/ImmutableStorage.java
@@ -6,23 +6,23 @@ import java.util.stream.IntStream;
 
 public class ImmutableStorage<T> implements Storage<T> {
 
-    private Map.Immutable<Long, Object> map;
+    private Map.Immutable<Integer, Object> map;
 
     public ImmutableStorage() {
         this(io.usethesource.capsule.Map.Immutable.of());
     }
 
-    private ImmutableStorage(Map.Immutable<Long, Object> map) {
+    private ImmutableStorage(Map.Immutable<Integer, Object> map) {
         this.map = map;
     }
 
     @Override
-    public T get(final long i) {
+    public T get(final int i) {
         return (T)this.map.get(i);
     }
 
     @Override
-    public Storage<T> set(final long i, final T x) {
+    public Storage<T> set(final int i, final T x) {
         return new ImmutableStorage<>(this.map.__put(i, x));
     }
